requirements.txt
streamlit>=1.32.0
azure-ai-inference>=1.0.0b9
azure-ai-documentintelligence>=1.0.0
azure-search-documents>=11.6.0
azure-storage-blob>=12.19.0
azure-identity>=1.15.0
python-dotenv>=1.0.0
tiktoken>=0.6.0
pandas>=2.0.0
pypdf>=4.0.0
Pillow>=10.0.0
